move to_16string overloads
